Adobe_1B/requirements.txt
pdfminer.six
transformers
torch